🔧 refactor(download.sh, sync.sh): improve script robustness and add more checks 🔧 refactor(vscode/install_extensions.sh): normalize extension names to lowercase 🔧 refactor(vscode/settings.json): remove git.enableCommitSigning setting 🔧 refactor(zsh/.metis.zsh): adjust PATH order and add conda setup 🔧 refactor(zsh/.p10k.zsh): change icons and add rust version 🔧 refactor(zsh/.zshrc): source tmux start script
The scripts have been refactored to improve their robustness by adding more checks before performing certain operations. This prevents errors from occurring when the expected conditions are not met. The vscode extension names have been normalized to lowercase for consistency. The git.enableCommitSigning setting has been removed from vscode settings as it's not necessary. The PATH order in the zsh/.metis.zsh file has been adjusted and the conda setup has been added. The icons in the zsh/.p10k.zsh file have been changed and the rust version has been added. The tmux start script is now sourced in the zsh/.zshrc file.

+++ download.sh
@@ -35,16 +35,22 @@ for dir in */ .*/; do
             cd $SCRIPT_DIR
             ;;
         "gnupg")
-            gpg --armor --export > "$SCRIPT_DIR/gnupg/$(hostname -s).asc"
+            mkdir -p "$HOME/.gnupg"
+            chmod -R u=rw,u+X,go= $HOME/.gnupg
+            if [ -n "$(gpg --list-secret-keys --keyid-format=long)" ]; then
+                gpg --armor --export > "$SCRIPT_DIR/gnupg/$(hostname -s).asc"
+            fi
             ;;
         "ssh")
-            cp "$HOME/.ssh/id_rsa.pub" "$SCRIPT_DIR/ssh/$(hostname -s).pub"
+            if [ -f "$HOME/.ssh/id_rsa" ]; then
+                cp "$HOME/.ssh/id_rsa.pub" "$SCRIPT_DIR/ssh/$(hostname -s).pub"
+            fi
             ;;
         "vim")
             fetch_file "https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim" "$SCRIPT_DIR/vim/autoload"
             ;;
         "vscode")
-            if command -v my_command >/dev/null 2>&1; then
+            if command -v code >/dev/null 2>&1; then
                 echo "#!/bin/bash" > $SCRIPT_DIR/vscode/install_extensions.sh
                 code --list-extensions | while read -r extension; do
                     echo "code --install-extension \"$extension\"" >> $SCRIPT_DIR/vscode/install_extensions.sh
@@ -56,8 +62,10 @@ for dir in */ .*/; do
             rm -rf $SCRIPT_DIR/zsh/ohmyzsh
             rm -rf $SCRIPT_DIR/zsh/.antigen
             mkdir -p $SCRIPT_DIR/zsh/.antigen
-            sh -c "RUNZSH=no ZSH=$SCRIPT_DIR/zsh/ohmyzsh $(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh) --keep-zshrc"
             fetch_file "https://raw.githubusercontent.com/zsh-users/antigen/master/bin/antigen.zsh" "$SCRIPT_DIR/zsh/.antigen"
+            if [[ "$SHELL" == "/bin/zsh" ]]; then
+                sh -c "RUNZSH=no ZSH=$SCRIPT_DIR/zsh/ohmyzsh $(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh) --keep-zshrc"
+            fi
             ;;
     esac
 done
